Don't require root access in the container
It is currently not possible to run commands into the container because
the ENTRYPOINT runs a `nix-shell` that tries to write to the store or
the nix db (even if all the needed derivations are already built).

This fixes it by running `nix-shell` on the drv file resulting from the
evaluation of the `shell.nix` file. In that case, `nix-shell` doesn't
try to write to the store or the database if the nedded derivations are
present (which is the case as they have been built in the Dockerfile).