Keep the correct STACK_PLATFORM_VARIANT
Stack has one separate global cache directory for each container.
This directory is determined using the `STACK_PLATFORM_VARIANT`
environment varable, which in case of `stack --docker` is set to an
identifier unique to the container.

The `buildStackProject` nix function sets this variable to `nix`.
A consequence of this is that `stack --nix` builds and `stack --docker`
builds using this image share the same cache directory, which can cause
some inconsistencies.

This PR restores the correct `STACK_PLATFORM_VARIANT` variable inside
the nix-shell to keep both separate.